mcp_mysql/pyproject.toml
[project]
name = "mcp-mysql"
version = "0.1.0"
description = "MCP server for MySQL database access"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.11.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
]
